+++ CHANGELOG.md
@@ -6,10 +6,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.1.0a9] - 2026-04-02
 
 ### Added
 
+- `bpod.is_ready` property indicating whether a compiled FSM is loaded and ready to run
+- `bpod.is_queued` property indicating whether an FSM is queued to run after the current
+  one
 - `bpod.get_data` for returning trial data after a state machine run.
 
 ### Changed
@@ -22,6 +25,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
     - `ReadThread` for serial communication with the Bpod
     - `EventThread` for handling and structuring the incoming data
     - `SoftcodeThread` for executing soft-codes
+
+### Fixed
+
+- reorganization of finalizers for better garbage collection.
 
 ## [0.1.0a8] - 2026-03-04